requirements.txt
pandas
ipykernel
pyarrow
pyyaml
python-dotenv
azure-identity
azure-storage-blob